chore: migrate test suite to TypeScript + bump @stonyx/utils
Convert all 8 test files under test/ from JavaScript to TypeScript, switching
from the dist-test/ compile pipeline to the events-style on-the-fly tsx runner.
Preserves the Sprint 43 maybeRegisterAppShutdown coverage from #52 and types
all sinon stubs, process stubs, and QUnit callbacks without any, @ts-ignore,
or skips.

Dep bumps:
- @stonyx/utils 0.2.3-beta.21 -> 0.2.3-beta.22

Note: @stonyx/logs was intended to bump to 1.0.1-beta.15, but that version
was published broken (missing dist/ folder in tarball). Staying on beta.14
until logs publishes a fixed beta.

Test script changes:
- Drop build:test + dist-test/ pipeline
- Use node --import tsx node_modules/qunit/bin/qunit.js (not .bin/qunit, which
  breaks pnpm's shell shim for tsx)
- Add typecheck script (tsc --noEmit on both configs)

Pre/post test counts: 44 pass / 0 fail, identical in both directions.

Closes #53
