+++ package.json
@@ -32,12 +32,12 @@
   },
   "scripts": {
     "build": "tsc",
-    "build:test": "tsc -p tsconfig.test.json",
+    "typecheck": "tsc --noEmit && tsc --noEmit -p tsconfig.test.json",
     "start": "node .",
-    "test": "pnpm build && pnpm build:test && qunit 'dist-test/test/**/*-test.js'",
+    "test": "node --import tsx node_modules/qunit/bin/qunit.js 'test/**/*-test.ts'",
     "error": "echo \"Error: no error handling\" && exit 1",
     "postinstall": "node scripts/postinstall.js",
-    "prepublishOnly": "npm test"
+    "prepublishOnly": "npm run build && npm test"
   },
   "repository": {
     "type": "git",
@@ -70,6 +70,7 @@
     "@types/sinon": "^21.0.1",
     "qunit": "^2.24.1",
     "sinon": "^21.0.0",
+    "tsx": "^4.21.0",
     "typescript": "^5.8.3"
   }
 }
